factor out common newlib functions
