+++ hal/src/photon/newlib_stubs.cpp
@@ -30,52 +30,4 @@
 
 extern "C" {
 
-/*
- * Implement C++ new/delete operators using the heap
- */
-
-void *operator new(size_t size)
-{
-	return malloc(size);
-}
-
-void *operator new[](size_t size)
-{
-	return malloc(size);
-}
-
-void operator delete(void *p)
-{
-	free(p);
-}
-
-void operator delete[](void *p)
-{
-	free(p);
-}
-
-
-/* Default implementation for call made to pure virtual function. */
-void __cxa_pure_virtual() {
-  PANIC(PureVirtualCall,"Call on pure virtual");
-  while (1);
-}
-
-/* Provide default implemenation for __cxa_guard_acquire() and
- * __cxa_guard_release(). Note: these must be revisited if a multitasking
- * OS is ported to this platform. */
-__extension__ typedef int __guard __attribute__((mode (__DI__)));
-int __cxa_guard_acquire(__guard *g) {return !*(char *)(g);};
-void __cxa_guard_release (__guard *g) {*(char *)g = 1;};
-void __cxa_guard_abort (__guard *) {};
-
-}
-
-namespace __gnu_cxx {
-
-void __verbose_terminate_handler()
-{
-  abort();
-}
-
 }
