+++ modules/electron/system-part1/src/newlib.cpp
@@ -1,0 +1,2 @@
+
+#include "../../../../hal/src/stm32/newlib.cpp"
